added base class for ObservableLists
